+++ src/thor/timedistancebssmatrix.cc
@@ -34,7 +34,10 @@ namespace valhalla {
 namespace thor {
 
 // Constructor with cost threshold.
-TimeDistanceBSSMatrix::TimeDistanceBSSMatrix() : settled_count_(0), current_cost_threshold_(0) {
+TimeDistanceBSSMatrix::TimeDistanceBSSMatrix(const boost::property_tree::ptree& config)
+    : settled_count_(0), current_cost_threshold_(0),
+      max_reserved_labels_count_(
+          config.get<uint32_t>("max_reserved_labels_count_matrix", kInitialEdgeLabelCountMatrix)) {
 }
 
 float TimeDistanceBSSMatrix::GetCostThreshold(const float max_matrix_distance) const {