+++ valhalla/sif/edgelabel.h
@@ -12,6 +12,7 @@
 namespace valhalla {
 namespace sif {
 
+constexpr uint32_t kInitialEdgeLabelCountMatrix = 4000000;
 constexpr uint32_t kInitialEdgeLabelCountBD = 1000000;
 constexpr uint32_t kInitialEdgeLabelCountBDMatrix = 2000000;